[front-design] refactor: 프론트 메인페이지 변경
