+++ front/src/features/counter/Counter.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react'
 
 // 커스텀 훅에
-import { useAppSelector, useAppDispatch } from '../../app/hooks'
+import { useAppSelector, useAppDispatch } from '../../hooks/exhook'
 import {
   decrement,
   increment,
